Sequentialize all requests and execute  them (non-parallel) in order as received.
