+++ packages/common/src/apiutil/logging.ts
@@ -54,6 +54,6 @@ export const expressLogger = pinoLogger.child({ type: "express" })
 export const dbLogger = pinoLogger.child({ type: "database" })
 export const traceLogger = pinoLogger.child({ type: "trace" })
 requestLogger.level = ServerConfig.getInstance().requestLog()
-traceLogger.level = ServerConfig.getInstance().requestLog()
+traceLogger.level = ServerConfig.getInstance().traceLog()
 expressLogger.level = ServerConfig.getInstance().expressLog()
 dbLogger.level = ServerConfig.getInstance().databaseLog()
